ordered_stream

Struct FromSortedStream

pub struct FromSortedStream<S> {
    pub stream: S,
}

An OrderedStream wrapper around a Stream.

Unlike FromStream, the items in the Stream are themselves ordered with no additional data.

Fields
stream: S
Implementations

impl<S> FromSortedStream<S>

pub fn new(stream: S) -> Self
where S: Stream, S::Item: Ord,

Create a new OrderedStream by applying a split_item closure to each element produced by the original stream.