Struct ordered_stream::FromStreamDirect

pub struct FromStreamDirect<S, F> { /* private fields */ }

An OrderedStream wrapper around a Stream.

This does not use any future or past knowledge of elements, and so is suitable if the stream rarely or never blocks. Prefer using FromStream if you plan to filter or join this stream and want other streams to be able to make progress while this one blocks.

Implementations

impl<S, F> FromStreamDirect<S, F>

pub fn new<Ordering, Data>(stream: S, split_item: F) -> Selfwhere
S: Stream,
F: FnMut(S::Item) -> (Ordering, Data),
Ordering: Ord,

Create a new OrderedStream by applying a split_item closure to each element produced by the original stream.

pub fn with_ordering<Ordering>(
stream: S,
get_ordering: F
) -> FromStreamDirect<S, impl FnMut(S::Item) -> (Ordering, S::Item)>where
S: Stream,
F: FnMut(&S::Item) -> Ordering,
Ordering: Ord,

Helper function to simplify the creation of a stream when you have a get_ordering function.