ordered_stream

Struct Join

pub struct Join<A, B>
where
    A: OrderedStream,
    B: OrderedStream<Data = A::Data, Ordering = A::Ordering>,
{ /* private fields */ }

A stream for the join function.

Implementations

impl<A, B> Join<A, B>
where A: OrderedStream, B: OrderedStream<Data = A::Data, Ordering = A::Ordering>,

pub fn into_inner(self) -> (A, B, Option<(A::Data, A::Ordering)>)

Split into the source streams.

This method returns the source streams along with any buffered item and its ordering.

pub fn stream_a(self: Pin<&mut Self>) -> Pin<&mut A>

Provide direct access to the underlying stream.

This may be useful if the stream provides APIs beyond OrderedStream. Note that the join itself may be buffering an item from this stream, so you should consult Self::peek_buffered and, if needed, Self::take_buffered before polling it directly.

pub fn stream_b(self: Pin<&mut Self>) -> Pin<&mut B>

Provide direct access to the underlying stream.

This may be useful if the stream provides APIs beyond OrderedStream. Note that the join itself may be buffering an item from this stream, so you should consult Self::peek_buffered and, if needed, Self::take_buffered before polling it directly.

pub fn peek_buffered( self: Pin<&mut Self>, ) -> (Option<(&mut A::Data, &A::Ordering)>, Option<(&mut B::Data, &B::Ordering)>)

Allow access to the buffered item, if any.

At most one of the two sides will be Some. The returned item is a candidate for being the next item returned by the joined stream, but it could not be returned by the most recent OrderedStream::poll_next_before call.

pub fn take_buffered(self: Pin<&mut Self>) -> Option<(A::Data, A::Ordering)>

Remove the buffered item, if one is present.

This does not poll either underlying stream. See Self::peek_buffered for details on why buffering exists.