ordered_stream
Trait FusedOrderedStream

pub trait FusedOrderedStream: OrderedStream {
    // Required method
    fn is_terminated(&self) -> bool;
}

An OrderedStream that tracks if the underlying stream should be polled.

Required Methods

fn is_terminated(&self) -> bool

Returns true if the stream should no longer be polled.